create player and tests

+++ database/migrations/2020_11_20_223949_create_players_table.php
@@ -16,7 +16,7 @@ class CreatePlayersTable extends Migration
         Schema::create('players', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('document')->unique();
+            $table->bigInteger('document')->unique();
             $table->integer('shirt_number');
             $table->foreignId('team_id')->constraints()->nullable();
             $table->timestamps();
